Changed the UtilitesFunctions to FailureHandler Added block on chromedriver version to 73 Cleaned up code for screen recording
