+++ src/main/java/utilities/FailureHandler.java
@@ -1,0 +1,135 @@
+package utilities;
+
+import cucumber.api.Scenario;
+import managers.Context;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.monte.media.Format;
+import org.monte.media.math.Rational;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.io.FileHandler;
+
+import java.awt.*;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+import static org.monte.media.FormatKeys.*;
+import static org.monte.media.VideoFormatKeys.*;
+
+public final class FailureHandler {
+
+    private static final Logger log = LogManager.getLogger(Logger.class.getName());
+    private final static String recordingVideoName = "Recorded";
+    private static SpecializedScreenRecorder screenRecorder;
+
+    public static void takePageSource(Scenario scenario) {
+        String pageSource = Context.driverManager.getDriver().getPageSource();
+
+        String timeToPrint = getCurrentTime();
+
+        Path path = Paths.get("pageSources/" + scenario.getName() + "_" + timeToPrint + ".html");
+
+        try {
+            Files.createFile(path);
+            Files.write(path, pageSource.getBytes());
+            scenario.embed(pageSource.getBytes(), "text/html");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    public static void takeScreenshot(Scenario scenario) {
+        String timeToPrint = getCurrentTime();
+
+        try {
+            File screenSource = ((TakesScreenshot) Context.driverManager.getDriver()).getScreenshotAs(OutputType.FILE);
+            FileHandler.copy(
+                    screenSource,
+                    new File("screenshots/" + scenario.getName() + "_" + timeToPrint + ".png")
+            );
+            byte[] screenSource2 = ((TakesScreenshot) Context.driverManager.getDriver()).getScreenshotAs(OutputType.BYTES);
+            scenario.embed(screenSource2, "image/png");
+        } catch (Exception e) {
+            log.error("Failed to take screenshot");
+            e.printStackTrace();
+        }
+    }
+
+    public static SpecializedScreenRecorder setUpScreenRecorder() {
+        GraphicsConfiguration gc = GraphicsEnvironment
+                .getLocalGraphicsEnvironment()
+                .getDefaultScreenDevice()
+                .getDefaultConfiguration();
+
+        Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
+        int width = screenSize.width;
+        int height = screenSize.height;
+
+        Rectangle captureSize = new Rectangle(0, 0, width, height);
+
+        try {
+            screenRecorder = new SpecializedScreenRecorder(gc, captureSize,
+                    new Format(MediaTypeKey, MediaType.FILE, MimeTypeKey, MIME_AVI),
+                    new Format(MediaTypeKey, MediaType.VIDEO, EncodingKey, ENCODING_AVI_TECHSMITH_SCREEN_CAPTURE,
+                            CompressorNameKey, ENCODING_AVI_TECHSMITH_SCREEN_CAPTURE,
+                            DepthKey, 24, FrameRateKey, Rational.valueOf(15),
+                            QualityKey, 1.0f,
+                            KeyFrameIntervalKey, 15 * 60),
+                    new Format(MediaTypeKey, MediaType.VIDEO, EncodingKey, "black",
+                            FrameRateKey, Rational.valueOf(30)),
+                    null, new File("videos/"), recordingVideoName);
+        } catch (IOException e) {
+            log.info("Failed to initialize video recording");
+            e.printStackTrace();
+        } catch (AWTException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    public static void startVideoRecord() {
+        try {
+            screenRecorder.start();
+        } catch (IOException e) {
+            log.info("Failed to start video recording");
+            e.printStackTrace();
+        }
+    }
+
+    public static void stopVideoRecord(Scenario scenario) {
+        try {
+            screenRecorder.stop();
+        } catch (IOException e) {
+            log.info("Failed to stop video recording");
+            e.printStackTrace();
+        }
+        renameVideoFile(scenario);
+    }
+
+    private static void renameVideoFile(Scenario scenario) {
+        String timeToPrint = getCurrentTime();
+        String directory = "videos/";
+
+        File oldFile = new File(directory + recordingVideoName + ".avi");
+        File newFile = new File(directory + scenario.getName() + "_" + timeToPrint + ".avi");
+
+        boolean success = oldFile.renameTo(newFile);
+
+        if (!success) {
+            log.error("Failed to save video file with correct name");
+        }
+    }
+
+    private static String getCurrentTime() {
+        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH.mm.ss");
+        LocalDateTime currentTime = LocalDateTime.now();
+        return dateTimeFormatter.format(currentTime);
+    }
+}
